v 1.0

+++ metaboxes/HighlightArticle.php
@@ -3,6 +3,7 @@
 class HighlightArticle
 {
     const HIGHLIGHT = 'catalogue_highlight';
+    const NONCE = 'catalogue_highlight_nonce';
 
     public static function register()
     {
@@ -18,15 +19,23 @@ class HighlightArticle
     public static function render($post)
     {
         $value = get_post_meta($post->ID, self::HIGHLIGHT, true);
+        wp_nonce_field(self::NONCE, self::NONCE)
         ?>
-        <input id="highlight" type="checkbox" value="1" name="<?= self::HIGHLIGHT ?>" <?= checked($value, '1') ?>>
-        <label for="highlight">Cet article est mise en avant</label>
+        <div class="cat-pad">
+        <span>
+        <input id="<?= self::HIGHLIGHT ?>"
+               type="checkbox" value="1"
+               name="<?= self::HIGHLIGHT ?>" <?= checked($value, '1') ?>>
+        </span>
+            <label class="components-checkbox-control__label" for="<?= self::HIGHLIGHT ?>">Cet article est mise en
+                avant</label>
+        </div>
         <?php
     }
 
     public static function save($post)
     {
-        if (array_key_exists(self::HIGHLIGHT, $_POST)) {
+        if (array_key_exists(self::HIGHLIGHT, $_POST) && wp_verify_nonce($_POST[self::NONCE], self::NONCE)) {
             if ($_POST[self::HIGHLIGHT] === '0') {
                 delete_post_meta($post, self::HIGHLIGHT);
             } else {
